Add initial handler for form in separate component
